feat: Add explicit schema bucket APIs (#34)
* feat: Add explicit schema bucket APIs

This commit introduces Cloud 2 specific APIs for managing
explicit schema buckets. Most of the changes are in the `cloud`
only paths for managing measurement schema.

The Bucket type adds a `schemaType` property, that is expected to
return `implicit` for OSS 2.x and is not modifiable. For Cloud 2
it can be either `implicit` or `explicit`, once the feature is
available.

* feat: Add name query parameter

This allows a client to find a measurement schema within a bucket by
name.