feat(unity): add query for org creation allowance
This introduces the concept of an "allowance": is the current user/org/account **allowed** to perform an operation? The response contains the answer to the question as well as additional information that allows a UI to show more useful information.

For now, all we need is a flag indicating whether a plan upgrade would allow an operation that is currently denied, but we can expand this in the future when/if it becomes necessary.

Implementation-wise, an allowance encapsulates any business logic around permitting an operation, including entitlements (aka plan limits) as well as (in the future) roles/permissions.